Step 6 — routes/admin.ts (881 lines) becomes routes/admin/ (7 files).
Cross-cutting concerns (authenticate + adminWriteLimiter) apply ONCE in
admin/index.ts before any sub-router runs; each sub-router declares
its own per-route permission gate. Banner comments marking the bolt-on
points (// ======= Reports / Notifications) are gone — those endpoints
now live in their own files.

New layout:
  routes/admin/index.ts          — composite mount, one error handler
  routes/admin/dashboard.ts      — /dashboard + /analytics (the cross-
                                   feature aggregate composer; pulls
                                   the topBookmarkedPosts slice from
                                   bookmarkService)
  routes/admin/comments.ts       — pending queue + per-id approve/reject
                                   + bulk-approve/bulk-reject (bulk body
                                   schema lifted out as a const)
  routes/admin/users.ts          — listing, ban/unban (with the moderator
                                   boundary checks intact), mention-mute
                                   toggle; local getUserGroupNames helper
                                   moves here with its only caller
  routes/admin/audit.ts          — single GET /audit endpoint
  routes/admin/reports.ts        — list + bulk-status; STATUS_VALUES
                                   tuple replaces the inline string array
  routes/admin/notifications.ts  — stats + send + broadcast

URL surface unchanged — every /api/v1/admin/* path responds exactly as
before.

Tested: bunx tsc --noEmit clean on server + client; bun test 45 pass.
